+++ xoto3/sqs/__init__.py
@@ -1,0 +1,1 @@
+from .utils import yield_and_ack_sqs_messages, yield_record_bodies_from_sqs_event  # noqa