+++ xoto3/utils/tree_map.py
@@ -118,6 +118,8 @@ def _tuple_starts_with(a: tuple, b: tuple) -> bool:
 
 
 def make_path_stop_transform(target_path: KeyPath, transform: SimpleTransform) -> PathTransform:
+    target_path = tuple(target_path)  # in case it's a list or something
+
     def path_tx(item: Any, path: KeyPath) -> Tuple[Any, bool]:
         stop = not _tuple_starts_with(target_path, path) or path == target_path
         # stop if not part of target path, or if we've reached the full path
@@ -129,6 +131,8 @@ def make_path_stop_transform(target_path: KeyPath, transform: SimpleTransform) -
 
 
 def make_path_only_transform(target_path: KeyPath, transform: SimpleTransform) -> PathTransform:
+    target_path = tuple(target_path)  # in case it's a list or something
+
     def path_only_tx(item: Any, path: KeyPath) -> Tuple[Any, bool]:
         if path == target_path:
             return transform(item), False